requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Google Drive API
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# PDF Processing
PyPDF2==3.0.1

# Data Processing
pandas==2.1.3
pyarrow==14.0.1

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Testing (optional)
pytest==7.4.3